인송님구현x..일단 합침

+++ Project/src/main/java/com/mvc/grade/Grade.java
@@ -25,7 +25,7 @@ public class Grade {
 	private Long listCode;
 	
 	@Column(nullable = true)
-	private Long userCode;
+	private int userCode;
 	
 	@Column(nullable = true)
 	private Double grade;
